feat(ui): rework compare toolbar
